fix(defcon): widen BIGINT end-to-end + writer/DDL parity guard (#211)
Cascade fix surfaced when PR #210's `monotonically_increasing_id().cast("long")`
finally let the DEFCON pipeline re-execute against fresh data. Three latent
schema lies fired in sequence on the first re-trigger:

1. `valued_schema.action_player_id = StringType()` vs pandas int64 →
   `pyarrow.lib.ArrowTypeError: Expected a string or bytes dtype, got int64`
   in applyInPandas.
2. `valued_schema.defcon_value = DoubleType()` vs bronze DDL FLOAT →
   `[DELTA_FAILED_TO_MERGE_FIELDS] defcon_value` on the replaceWhere write.
3. `stg_defcon__results.cast(... as int)` on five BIGINT-bearing ID columns →
   `[CAST_OVERFLOW] LONG → INT` on the not_null staging tests.

Each had been hidden by SkipGuard for many cycles. Fix:

- defcon_lite_360.py + defcon_lite_tracking.py: action_player_id LongType,
  defcon_value FloatType in both credits_schema and valued_schema.
- stg_defcon__results.sql: 5 ID columns cast as bigint, not int
  (competition_id, season_id, defender_player_id, defender_team_id,
  action_player_id).
- _marts__models.yml + fct_defcon_actions/fct_defcon_pressure/fct_defensive_values:
  12 contract data_types int → bigint across the three defcon marts; matching
  cast(null as bigint) updates to the {% else %} disabled-state placeholders;
  test_marts_live_schema.py expected types updated to match.

Defense in depth (ADR-002 §4 confirming application):
- src/tests/test_defcon_schema_parity.py: parses _RESULTS_SCHEMA DDL string
  and asserts column-name + Spark-type equality with valued_schema for both
  defcon_lite_360 and defcon_lite_tracking, plus cross-module DDL equality.
  Catches all four of the above drift classes at CI time.

Live-invariant calibration (test_marts_kimball_contracts.py):
- Removed the two defender player_key threshold tests on fct_defensive_values
  and fct_defcon_actions. The previous 16% structural floor reported in
  Phase 0 Task 0.7 was an artifact of the broken INT cast truncating
  `monotonically_increasing_id()` LONG values into the real player_id range
  by accident; the post-fix structural floor is 0.06% (469 / 828k), which
  is correct — synthetic 360-anonymous defenders should not resolve to real
  dim_players entries. The test cannot meaningfully regression-check on
  noise.
- Raised team_key threshold 0.10 → 0.99 on the same two marts. PR #209's
  teammate-flag derivation now resolves at 100% (828634/828634) on dev_gold.

End-to-end verification on dev_gold (compute_defcon_lite run_id=163853960124728,
dbt_build run_id=559814695983073, synced refresh, live invariants):

  total rows                        828,634
  match_key set                       100.0%
  team_key set                        100.0%
  action_player_key set               100.0%
  player_key (defender, synthetic)      0.06%   (expected — not tested)
  fct_defcon_actions_synced              COMPLETE
  fct_defcon_pressure_synced             COMPLETE
  fct_defensive_values_synced            COMPLETE
  test_marts_kimball_contracts.py       29/29 PASS
  test_marts_live_schema.py             8/8 PASS
  test_idsse_is_progressive_coverage    2/2 PASS

Wheel: 0.3.15 → 0.3.18.

Documentation:
- ADR-002 §4 now lists DEFCON valued_schema as a confirming application of
  the writer/target schema drift guard pattern.
- ADR-011 staged-rollout table moves PR 5 / PR 6 from Planned → Shipped
  with PR numbers and followup chain.
- TODO.md last-updated → 2026-04-27 with the full followup sequence.

+++ dbt_project/models/staging/defcon/stg_defcon__results.sql
@@ -26,16 +26,26 @@ deduplicated as (
 
 cleaned as (
 
+    -- All five integer-bearing columns are cast to BIGINT to match the
+    -- bronze.defcon_results contract (`_RESULTS_SCHEMA` declares BIGINT
+    -- for competition_id, season_id, defender_player_id, defender_team_id,
+    -- and action_player_id). The previous `as int` casts were latent
+    -- CAST_OVERFLOW bombs that detonated 2026-04-27 once PRs #209/#210
+    -- and the StringType→LongType followup widened bronze ID columns to
+    -- their declared BIGINT types — `monotonically_increasing_id()`
+    -- returns 64-bit values and Spark ANSI rejects LONG→INT statically.
+    -- defcon_value is widened FLOAT→DOUBLE (lossless) for downstream
+    -- arithmetic; bronze remains FLOAT per the writer schema parity test.
     select
         cast(event_id as string)                as event_id,
         cast(match_id as string)                as match_id,
-        cast(competition_id as int)             as competition_id,
-        cast(season_id as int)                  as season_id,
-        cast(defender_player_id as int)         as defender_player_id,
-        cast(defender_team_id as int)           as defender_team_id,
+        cast(competition_id as bigint)          as competition_id,
+        cast(season_id as bigint)               as season_id,
+        cast(defender_player_id as bigint)      as defender_player_id,
+        cast(defender_team_id as bigint)        as defender_team_id,
         cast(defender_x as double)              as defender_x,
         cast(defender_y as double)              as defender_y,
-        cast(action_player_id as int)           as action_player_id,
+        cast(action_player_id as bigint)        as action_player_id,
         cast(action_type as string)             as action_type,
         cast(action_x as double)               as action_x,
         cast(action_y as double)               as action_y,
